Clean up '-Wint-to-pointer-cast' warning.

+++ test/cgi_task_queue_test.c
@@ -5,14 +5,14 @@
 
 static void* print(void *arg)
 {
-	printf("%d\n", (int)arg);
+	printf("%ld\n", (long)arg);
 	return NULL;
 }
 
 int main()
 {
 	cgi_task_queue_t *head = NULL;
-	int i;
+	long i;
 	for(i = 0; i < 10; ++i)
 	{
 		cgi_task_queue_push(&head, print, (void*)i);
